Turn ContentRenderer into exportable KContentRenderer.
