+++ kolibri/core/assets/src/core-app/index.js
@@ -1,3 +1,7 @@
+/**
+ * Provides the public API for the Kolibri FrontEnd core app.
+ * @module Facade
+ */
 // polyfill for older browsers
 // TODO: rtibbles whittle down these polyfills to only what is needed for the application
 import 'core-js';
@@ -12,16 +16,43 @@ import trackInputModality from 'kolibri-components/src/styles/trackInputModality
 import keenUiConfig from 'keen-ui/src/config';
 import branding from 'kolibri.utils.branding';
 import logging from 'kolibri.lib.logging';
+import Vue from 'vue';
+import VueMeta from 'vue-meta';
+import VueRouter from 'vue-router';
+import Vuex from 'vuex';
+import KThemePlugin from 'kolibri-components/src/KThemePlugin';
+import heartbeat from 'kolibri.heartbeat';
+import KContentPlugin from 'kolibri-components/src/content/KContentPlugin';
 import keenOptions from '../keen-config/options.json';
-import CoreAppConstructor from './constructor';
+import { i18nSetup, languageDirection } from '../utils/i18n';
+import ContentRendererLoadingComponent from '../views/ContentRenderer/ContentRendererLoading';
+import ContentRendererErrorComponent from '../views/ContentRenderer/ContentRendererError';
+import apiSpec from './apiSpec';
 import monitorPageVisibility from './monitorPageVisibility';
 import getPluginData from 'kolibri.utils.getPluginData';
 // Do this before any async imports to ensure that public paths
 // are set correctly
 urls.setUp();
 
+// Shim window.location.origin for IE.
+if (!window.location.origin) {
+  window.location.origin = `${window.location.protocol}//${window.location.hostname}${
+    window.location.port ? `:${window.location.port}` : ''
+  }`;
+}
+
 // set up logging
 logging.setDefaultLevel(process.env.NODE_ENV === 'production' ? 2 : 0);
+
+/**
+ * Object that forms the public API for the Kolibri
+ * core app.
+ */
+const coreApp = {
+  // Assign API spec
+  ...apiSpec,
+  version: __version,
+};
 
 // set up theme
 const kolibriTheme = getPluginData().kolibriTheme;
@@ -43,10 +74,27 @@ trackInputModality();
 // monitor page visibility
 monitorPageVisibility();
 
-// Create an instance of the global app object.
+// Register Vue plugins and components
+Vue.use(Vuex);
+Vue.use(VueRouter);
+Vue.use(VueMeta);
+Vue.use(KThemePlugin);
+
+Vue.use(KContentPlugin, {
+  languageDirection,
+  ContentRendererLoadingComponent,
+  ContentRendererErrorComponent,
+  facade: coreApp,
+  logging: logging.getLogger(`${__kolibriModuleName} plugin mediator`),
+  activeCallback: heartbeat.setActive,
+});
+
+// Start the heartbeat polling here, as any URL needs should be set by now
+heartbeat.startPolling();
+
+i18nSetup().then(coreApp.ready);
+
 // This is exported by webpack as the kolibriCoreAppGlobal object, due to the 'output.library' flag
 // which exports the coreApp at the bottom of this file as a named global variable:
 // https://webpack.github.io/docs/configuration.html#output-library
-const coreApp = new CoreAppConstructor();
-
 export default coreApp;
