+++ kolibri/core/assets/src/utils/i18n.js
@@ -1,10 +1,17 @@
 import FontFaceObserver from 'fontfaceobserver';
 import vue from 'kolibri.lib.vue';
 import logger from 'kolibri.lib.logging';
+import { languageDirections, defaultLanguage } from 'kolibri-components/src/utils/i18n';
 import importIntlLocale from './intl-locale-data';
 import importVueIntlLocaleData from './vue-intl-locale-data';
 import getPluginData from 'kolibri.utils.getPluginData';
 
+export {
+  languageDirections,
+  defaultLanguage,
+  languageValidator,
+  getContentLangDir,
+} from 'kolibri-components/src/utils/i18n';
 export { licenseTranslations } from './licenseTranslations';
 
 const logging = logger.getLogger(__filename);
@@ -27,35 +34,16 @@ function $trWrapper(nameSpace, defaultMessages, formatter, messageId, args) {
   return formatter(message, args);
 }
 
-export const languageDirections = {
-  LTR: 'ltr',
-  RTL: 'rtl',
-};
-
-const defaultLocale = 'en';
-
-export const defaultLanguage = {
-  id: 'en',
-  lang_name: 'English',
-  lang_direction: languageDirections.LTR,
-};
-
-export const languageValidator = language => {
-  return ['id', 'lang_name', 'lang_direction'].reduce((valid, key) => valid && language[key], true);
-};
+const defaultLocale = defaultLanguage.id;
 
 export const availableLanguages = {
-  en: defaultLanguage,
+  [defaultLocale]: defaultLanguage,
 };
 
 export let currentLanguage = defaultLocale;
 
 // Default to ltr
 export let languageDirection = languageDirections.LTR;
-
-export const getContentLangDir = language => {
-  return (language || {}).lang_direction || languageDirections.LTR;
-};
 
 export const getLangDir = id => {
   return (availableLanguages[id] || {}).lang_direction || languageDirections.LTR;
