+++ packages/kolibri-components/src/content/constants.js
@@ -1,0 +1,1 @@
+export const RENDERER_SUFFIX = '_renderer';
